Add tmux config for docker container
Loosely based on
https://github.com/js-jslog/dotfiles/commit/0549d974a6a67e13e42daf96861ca0f9f66d40e0
but it's been a long time since I used that to be fair and I have
simplified it.

This is the context that I'm working in now so it's all bundled
together. If I get another tmux context which requires mostly the same
but a bit different then I'll pull apart the common items.